Get Contact Form Inputs
Had to figure out how to get the inputs from my contact form; changes to both index.htm and app.py to successfully do this and return the user to the correct anchor in the contact form.

If statements can be embedded into the HTML to display an optional thank you message after the user hits the submit button.

Script tags can be used to run a script in the HTML page.

Variables can be passed to an HTML page, such as this submit_button_pressed variable here where contact is the value. This can then be used in the if statement to display the thank you message.