Partition Key Update
It turns out shards/partitions are enforced, and you have to manually add partition keys into the data itself. The format for this is:

```
{'id': email, 'name': name, 'subject': subject, 'message': message, 'portfolio_section': "contact_form"}
```

Where portfolio_section in this case is considered the partition key. This key value is necessary when doing point read/writes, as opposed to doing general multi-partition SQL queries which are much more expensive.