Create Utilities File, Disc Drive and Accelerator Project HTML Files
I separated the projects themselves into individual HTML files so that I could customize those pages individually and not take up room on the home page. In addition, the images are clickable in Index.htm to serve as the links. On the project pages, I added a return button that will take you back to the projects section of the home page. The Utilities.py file will be used to hold commonly recurring chunks of code that can be useful in multiple places, and also cleans up some of app.py.