Final Update for 503 Site Unavailable Error Fix
Fixed the 503 Site Unavailable Error finally. It turned out to be an issue with the Microsoft Azure Cloud Service's workers. The free account did not have any workers available, so I had to upgrade the portfolio_app to B1 service plan to get a worker, and then I can downgrade again to get back to the free service.